Add. de AutoMapper na Camada Aplications

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,9 +1,9 @@
-﻿using System;
+using System;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ProEventos.Application.Contratos;
-using ProEventos.Domain;
+using ProEventos.Application.Dtos;
 // ====================================
 
 namespace ProEventos.API.Controllers
@@ -24,7 +24,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado !");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -42,7 +42,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Evento(s) por Código(Id) não encontrado(s) !");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -60,7 +60,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (evento == null) return NotFound("Evento(s) por tema não encontrado(s) !");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -73,12 +73,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.AddEventos(model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar evento !");
+                if (evento == null) NoContent();
 
                 return Ok(evento);
             }
@@ -91,12 +91,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{Id}")]
-        public async Task <IActionResult> Put(int Id, Evento model)
+        public async Task <IActionResult> Put(int Id, EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.UpdateEventos(Id, model);
-                if (evento == null) return BadRequest("Erro ao tentar alterar evento !");
+                if (evento == null) NoContent();
 
                 return Ok(evento);
             }
@@ -113,9 +113,13 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+            var evento = await _eventoService.GetEventoByIdAsync(Id, true);
+            if (evento == null) NoContent();
+
+
              return await _eventoService.DeleteEventos(Id) ? 
                     Ok("Deletado com sucesso !") : 
-                    BadRequest("Evento não pode ser deletado !");            
+                    throw new Exception("Ocorreu um problema não especifico ao tentar deletar Evento !");       
                                    
             }
             catch (Exception ex)
